優化加載速度,移除不必要的嵌套佈局

+++ app/src/main/java/com/citrus/mCitrusTablet/view/adapter/OrderDeliveryAdapter.kt
@@ -45,7 +45,7 @@ class OrderDeliveryAdapter :RecyclerView.Adapter<OrderDeliveryAdapter.TasksViewH
                 flavorNameBlock.visibility =  View.GONE
                 addNameBlock.visibility =  View.GONE
                 gName.text = (position+1).toString()+". " + item.gname
-                price.text = "$"+item.price.toInt().toString()
+                price.text = "$"+item.price.toString()
                 qty.text = "x" + item.qty
 
                 if(item.addName.isNotEmpty() && item.addName != ""){
